refactor: extract useRelationshipSearch composable (graph state + onFindRelationships)
